+++ src/main/java/com/mraof/minestuck/data/MinestuckEnUsLanguageProvider.java
@@ -705,6 +705,7 @@ public class MinestuckEnUsLanguageProvider extends MinestuckLanguageProvider
 		addItem(MSItems.SHADOWRAZOR, "Shadowrazor");
 		addItem(MSItems.PRINCESS_PERIL, "Princess Peril");
 		addItemTooltip(MSItems.PRINCESS_PERIL, "Truly a dagger fit for a princess.");
+		addItem(MSItems.CHEESE_KNIFE, "Cheese Knife");
 		addItem(MSItems.HOUSE_KEY, "House Key");
 		addItemTooltip(MSItems.HOUSE_KEY, "This incredible item has the ability to open any wooden door! Wow.");
 		addItem(MSItems.KEYBLADE, "Keyblade");
@@ -1935,6 +1936,7 @@ public class MinestuckEnUsLanguageProvider extends MinestuckLanguageProvider
 		add(GristCommand.FAILURE, "Failed to modify the grist cache for %s.");
 		add(GristCommand.SET, "Set the grist cache for %s players to %s.");
 		add(SetRungCommand.SUCCESS, "Successfully changed the echeladder of %s players to rung %d with %d%% progress.");
+		add(GetRungCommand.SUCCESS, "%s's current echeladder rung is: %d.");
 		add(PorkhollowCommand.SEND, "Successfully sent %s boondollars to %s.");
 		add(PorkhollowCommand.RECEIVE, "Received %s boondollars from %s.");
 		add(PorkhollowCommand.TAKE, "Successfully took out %s boondollars from your porkhollow.");